requirements.txt
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
gevent==1.1rc1
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
greenlet==0.4.0
wsgiref==0.1.2
